separate chain specific functions
